refractoring, organized code a lot!

+++ Aux/ft_strlen.c
@@ -3,17 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dsa-mora <dsa-mora@student.42.fr>          +#+  +:+       +#+        */
+/*   By: duarte33 <duarte33@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/22 21:39:36 by dsa-mora          #+#    #+#             */
-/*   Updated: 2022/10/22 21:39:36 by dsa-mora         ###   ########.fr       */
+/*   Updated: 2023/09/10 18:40:06 by duarte33         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minirt.h"
 
 //Retorna o numero de elementos de uma str sem contar com o NULL
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	i;
 
@@ -26,4 +25,4 @@ size_t	ft_strlen(const char *str)
 // int main(void)
 // {
 //     printf("%i", ft_strlen("duarte"));
-// } 
+// }
